refactor: separate `yarnInstall` function into `src/yarn/install.ts` file
